inverse kinematics, claw angle

+++ TeamCode/src/main/java/org/firstinspires/ftc/teamcode/common/subsystem/ExtensionSubsystem.java
@@ -99,7 +99,7 @@ public class ExtensionSubsystem extends WSubsystem {
 
     }
 
-    public double getBackdropHeight() {
+    public int getBackdropHeight() {
         return backdropHeight;
     }
